Separated out the Package Manifest class to its own file
